requirements.txt
django
djangorestframework
django-oauth-toolkit
django-bootstrap3
